Allow float32 inputs in ALS models
Previously all the ALS based models expected sparse matrices with a dtype of
numpy.float64 - and would fail to run in cython code if passed a float32
matrix instead. Fix this to allow passing either matrix type, and default
to float32 since the extra precision won't help in the majority of cases.

+++ examples/lastfm.py
@@ -69,7 +69,7 @@ def read_data(filename):
     data['artist'] = data['artist'].astype("category")
 
     # create a sparse matrix of all the users/plays
-    plays = coo_matrix((data['plays'].astype(float),
+    plays = coo_matrix((data['plays'].astype(numpy.float32),
                        (data['artist'].cat.codes.copy(),
                         data['user'].cat.codes.copy())))
